requirements.txt
requests
statsapi
kafka-python
streamlit 
pandas
numpy
matplotlib
scikit-learn
xgboost
joblib